bridge/requirements.txt
pyserial==3.5
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
